requirements.txt
aiohttp==3.3.2
async-timeout==3.0.0
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
idna==2.7
idna-ssl==1.0.1
multidict==4.3.1
python-engineio==2.1.1
python-socketio==1.9.0
six==1.11.0
wincertstore==0.2
yarl==1.2.6
